requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==1.0.0asgiref==3.4.1astunparse==1.6.3cachetools==4.2.4certifi==2021.10.8charset-normalizer==2.0.9Django==4.0django-cors-headers==3.10.1djangorestframework==3.13.1flatbuffers==2.0gast==0.4.0google-auth==2.3.3google-auth-oauthlib==0.4.6google-pasta==0.2.0grpcio==1.43.0h5py==3.6.0idna==3.3importlib-metadata==4.10.0keras==2.7.0Keras-Preprocessing==1.1.2libclang==12.0.0Markdown==3.3.6numpy==1.21.5oauthlib==3.1.1opencv-python==4.5.4.60opt-einsum==3.3.0Pillow==8.4.0protobuf==3.19.1pyasn1==0.4.8pyasn1-modules==0.2.8pydicom==2.2.2pytz==2021.3requests==2.26.0requests-oauthlib==1.3.0rsa==4.8six==1.16.0sqlparse==0.4.2tensorboard==2.7.0tensorboard-data-server==0.6.1tensorboard-plugin-wit==1.8.0tensorflow==2.7.0tensorflow-addons==0.15.0tensorflow-estimator==2.7.0tensorflow-io-gcs-filesystem==0.23.1termcolor==1.1.0typeguard==2.13.3typing_extensions==4.0.1tzdata==2021.5urllib3==1.26.7Werkzeug==2.0.2wrapt==1.13.3zipp==3.6.0